Added GameLoop package which will hold the different "areas" of our game (i.e. At sea, in port, etc.).
Created an enum and method to call these areas from the main run time.

Added TimeTracker singleton.

Updated Player to a singleton design.

Updated Opening Loop to ask for and store the user name.  Mostly to prove out the singleon design of the player.

Made a few tweaks to ships and attributes.

Updated 4/12 11:11 AM